refactor(web): split member.tsx into hooks and components
member.tsx had grown to 1235 lines holding the chat, the composer, patient
controls, referrals and checklists. Split it along the boundaries main already
established with ChecklistWidget/useChecklists/CollapsibleSection: hooks own the
Firestore subscriptions, components own their own UI and handlers.

  useChat.ts            message state, live listener, pagination, read receipts
  usePatientProfile.ts  patient doc subscription, current user, social workers
  ChatPanel.tsx         chat column
  MessageBubble.tsx     one message + chunk grouping
  ChatComposer.tsx      textarea, image picker, send
  PatientControls.tsx   assignment / status / verify-deny
  format_datetime.ts    shared message timestamp + person name formatters

member.tsx is now a 156-line layout shell.

Strictly behavior-preserving: no bug fixes and no cleanups. In particular the
live-listener bug that discards older paginated pages moves across unchanged,
with a comment pointing at #77 — fixing it here would have made this diff
unreviewable.

The ad-hoc date formatters in _index.tsx and calendar.tsx were left alone
rather than folded into format_datetime.ts; their output differs (year vs no
year, time vs no time), so unifying them would change what renders.
